separate postgis to new module
